Don't fail validation when no RuntimesConfig
This relaxation allows projects to be read (including project.yml) when
there are runtime decls in the config but the actual RuntimesConfig
wasn't read.  There are valid cases for this when the caller has no
credential store and no API host in the environment but is just
inspecting a project.  It is true that invalid runtimes may be
accepted in that use-case (only).