Also drop 12.x from automated workflows
The new esbuild may be incompatible with 12.x.  But, I don't see why we
need to support it any more.  We can update the engine declarations
later.  There are very few direct incorporations of the deployer so as
long as 'nim' continues to work (which will be tested sooner or later)
we are ok.